feat : started order managment

+++ products/models.py
@@ -37,7 +37,7 @@ class Product(models.Model):
     name = models.CharField(max_length=100)
     slug = models.SlugField(max_length=200,unique=True, blank=True)
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name='products')
-    price = models.DecimalField(default=0, decimal_places=2, max_digits=10)
+    price = models.PositiveIntegerField(default=0)
     description = models.CharField(max_length=500)
     image = models.ImageField(upload_to='uploads/product')
     stock = models.PositiveIntegerField(default=0)
